Update CfgVehicles
Reduced amount of CfgVehicles files by merging opfor and independent into one
