--- addons/msf/CfgVehicles_O_Desert.hpp
@@ -1,49 +1,0 @@
-class TACU_MSF_U_O_Desert_Rifleman: TACU_MSF_U_I_Desert_Rifleman {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_Rifleman_AT: TACU_MSF_U_I_Desert_Rifleman_AT {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_Autorifleman: TACU_MSF_U_I_Desert_Autorifleman {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_TeamLeader: TACU_MSF_U_I_Desert_TeamLeader {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_Engineer: TACU_MSF_U_I_Desert_Engineer {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_SMG: TACU_MSF_U_I_Desert_SMG {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_Medic: TACU_MSF_U_I_Desert_Medic {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_Grenadier: TACU_MSF_U_I_Desert_Grenadier {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_Marksman: TACU_MSF_U_I_Desert_Marksman {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Desert_Crewman: TACU_MSF_U_I_Desert_Crewman {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
