--- addons/msf/CfgVehicles_O_Wood.hpp
@@ -1,49 +1,0 @@
-class TACU_MSF_U_O_Wood_Rifleman: TACU_MSF_U_I_Wood_Rifleman {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_Rifleman_AT: TACU_MSF_U_I_Wood_Rifleman_AT {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_Autorifleman: TACU_MSF_U_I_Wood_Autorifleman {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_TeamLeader: TACU_MSF_U_I_Wood_TeamLeader {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_Engineer: TACU_MSF_U_I_Wood_Engineer {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_SMG: TACU_MSF_U_I_Wood_SMG {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_Medic: TACU_MSF_U_I_Wood_Medic {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_Grenadier: TACU_MSF_U_I_Wood_Grenadier {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_Marksman: TACU_MSF_U_I_Wood_Marksman {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
-
-class TACU_MSF_U_O_Wood_Crewman: TACU_MSF_U_I_Wood_Crewman {
-    side = 0;
-    faction = "TACU_MSF_O";
-};
