+++ addons/msf/CfgVehicles_Wood.hpp
@@ -7,6 +7,11 @@ class TACU_MSF_U_I_Wood_Rifleman: TACU_MSF_U_I_Desert_Rifleman {
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
 };
 
+class TACU_MSF_U_O_Wood_Rifleman: TACU_MSF_U_I_Wood_Rifleman {
+    side = 0;
+    faction = "TACU_MSF_O";
+};
+
 class TACU_MSF_U_I_Wood_Rifleman_AT: TACU_MSF_U_I_Desert_Rifleman_AT {
     uniformClass = "TACU_MSF_Uniform_Combat_TigerWood";
     backpack = "TACU_MSF_B_AT_Green";
@@ -14,6 +19,11 @@ class TACU_MSF_U_I_Wood_Rifleman_AT: TACU_MSF_U_I_Desert_Rifleman_AT {
     linkedItems[] = {DEFAULT_ITEMS_RADIO, "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Engineer"};
     respawnLinkedItems[] = {DEFAULT_ITEMS_RADIO, "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Engineer"};
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
+};
+
+class TACU_MSF_U_O_Wood_Rifleman_AT: TACU_MSF_U_I_Wood_Rifleman_AT {
+    side = 0;
+    faction = "TACU_MSF_O";
 };
 
 class TACU_MSF_U_I_Wood_Autorifleman: TACU_MSF_U_I_Desert_Autorifleman {
@@ -24,12 +34,22 @@ class TACU_MSF_U_I_Wood_Autorifleman: TACU_MSF_U_I_Desert_Autorifleman {
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
 };
 
+class TACU_MSF_U_O_Wood_Autorifleman: TACU_MSF_U_I_Wood_Autorifleman {
+    side = 0;
+    faction = "TACU_MSF_O";
+};
+
 class TACU_MSF_U_I_Wood_TeamLeader: TACU_MSF_U_I_Desert_TeamLeader {
     uniformClass = "TACU_MSF_Uniform_Combat_TigerWood";
     //editorPreview = QPATHTOF(ui\MSF_U_I_Wood_TeamLeader.jpg);
     linkedItems[] = {DEFAULT_ITEMS_RADIO, "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Officer"};
     respawnLinkedItems[] = {DEFAULT_ITEMS_RADIO, "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Officer"};
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
+};
+
+class TACU_MSF_U_O_Wood_TeamLeader: TACU_MSF_U_I_Wood_TeamLeader {
+    side = 0;
+    faction = "TACU_MSF_O";
 };
 
 class TACU_MSF_U_I_Wood_Engineer: TACU_MSF_U_I_Desert_Engineer {
@@ -41,12 +61,22 @@ class TACU_MSF_U_I_Wood_Engineer: TACU_MSF_U_I_Desert_Engineer {
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
 };
 
+class TACU_MSF_U_O_Wood_Engineer: TACU_MSF_U_I_Wood_Engineer {
+    side = 0;
+    faction = "TACU_MSF_O";
+};
+
 class TACU_MSF_U_I_Wood_SMG: TACU_MSF_U_I_Desert_SMG {
     uniformClass = "TACU_MSF_Uniform_Combat_TigerWood";
     //editorPreview = QPATHTOF(ui\MSF_U_I_Wood_SMG.jpg);
     linkedItems[] = {DEFAULT_ITEMS_RADIO, "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Scout"};
     respawnLinkedItems[] = {DEFAULT_ITEMS_RADIO, "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Scout"};
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
+};
+
+class TACU_MSF_U_O_Wood_SMG: TACU_MSF_U_I_Wood_SMG {
+    side = 0;
+    faction = "TACU_MSF_O";
 };
 
 class TACU_MSF_U_I_Wood_Medic: TACU_MSF_U_I_Desert_Medic {
@@ -58,12 +88,22 @@ class TACU_MSF_U_I_Wood_Medic: TACU_MSF_U_I_Desert_Medic {
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
 };
 
+class TACU_MSF_U_O_Wood_Medic: TACU_MSF_U_I_Wood_Medic {
+    side = 0;
+    faction = "TACU_MSF_O";
+};
+
 class TACU_MSF_U_I_Wood_Grenadier: TACU_MSF_U_I_Desert_Grenadier {
     uniformClass = "TACU_MSF_Uniform_Combat_TigerWood";
     //editorPreview = QPATHTOF(ui\MSF_U_I_Wood_Grenadier.jpg);
     linkedItems[] = {DEFAULT_ITEMS_RADIO, "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Grenadier"};
     respawnLinkedItems[] = {DEFAULT_ITEMS_RADIO, "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Grenadier"};
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
+};
+
+class TACU_MSF_U_O_Wood_Grenadier: TACU_MSF_U_I_Wood_Grenadier {
+    side = 0;
+    faction = "TACU_MSF_O";
 };
 
 class TACU_MSF_U_I_Wood_Marksman: TACU_MSF_U_I_Desert_Marksman {
@@ -74,10 +114,7 @@ class TACU_MSF_U_I_Wood_Marksman: TACU_MSF_U_I_Desert_Marksman {
     editorSubcategory = "TACU_MSF_EdSubCat_Wood";
 };
 
-class TACU_MSF_U_I_Wood_Crewman: TACU_MSF_U_I_Desert_Crewman {
-    uniformClass = "TACU_MSF_Uniform_Combat_TigerWood";
-    //editorPreview = QPATHTOF(ui\MSF_U_I_Wood_Crewman.jpg);
-    linkedItems[] = {DEFAULT_ITEMS_RADIO, "H_HelmetCrew_B", "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Crewman"};
-    respawnLinkedItems[] = {DEFAULT_ITEMS_RADIO, "H_HelmetCrew_B", "G_Balaclava_TI_blk_F", "CUP_V_B_BAF_DPM_Osprey_Mk3_Crewman"};
-    editorSubcategory = "TACU_MSF_EdSubCat_Wood";
+class TACU_MSF_U_O_Wood_Marksman: TACU_MSF_U_I_Wood_Marksman {
+    side = 0;
+    faction = "TACU_MSF_O";
 };
